Add is_ingredient function to NLU, improve get_ingredients.
is_ingredient tries to use WordNet to determine if a word is an ingredient.  It
falls back on checking the ingredients wordlist.  I modified get_ingredients to
use is_ingredient, so this should reduce the number of false misunderstandings
when searching small databases that might not have any recipes with a certain
ingredient.